bcmdhd: Enable support for new cfg80211_port_authorized format
In order to support Transition disable policy when 4way HS is
offloaded a new format of cfg80211_port_authorized was
supported with additonal parameters to indicate transition
disable policy. The support for the new cfg80211_port_authorized
format was disabled under the flag WL_MLO_BKPORT_NEW_PORT_AUTH.
In the current change, the flag is enabled in the build config.

Bug: 236661101
Test: Verified kernel compilation.
Change-Id: I131841ed3189234b0d06da8605f1521dbccdaf9f
Signed-off-by: kensun <kensun@google.com>